Move evals to own submodule
